tests/requirements-ci.txt
setuptools==80.0.0
pytest-html
coverage
pytest-env
requests
pytest-xdist
openpyxl
pytest-repeat